Refactor approval infrastructure
More pull request approval infrastructure for re-use for `pull_request_review` payloads.
